refactor: Update seat_price and base_price to use integer data type

+++ db/migrate/20240721165557_create_flights.rb
@@ -4,7 +4,7 @@ class CreateFlights < ActiveRecord::Migration[6.1]
 
       t.string :name, null: false
       t.integer :no_of_seats
-      t.decimal :base_price, null: false
+      t.integer :base_price, null: false
       t.datetime :departs_at, null: false
       t.datetime :arrives_at, null: false
