+++ db/migrate/20240721165713_create_bookings.rb
@@ -3,7 +3,7 @@ class CreateBookings < ActiveRecord::Migration[6.1]
     create_table :bookings do |t|
 
       t.integer :no_of_seats, null: false
-      t.decimal :seat_price, null: false
+      t.integer :seat_price, null: false
       t.belongs_to :user, null: false, foreign_key: true, index: true
       t.belongs_to :flight, null: false, foreign_key: true, index: true
